django/requirements.txt
# Core
Django==1.10.8
wheel==0.24.0

# Rest apis
djangorestframework==3.5.4

# Testing
factory-boy==2.11.1
django-nose==1.4.5
